requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20
streamlit>=1.48.0
requests>=2.32.4
pypdf>=5.9.0
pydantic-settings>=2.7.1
python-dotenv>=1.0.1
strands-agents>=1.25.0
openai>=1.61.0
anthropic>=0.42.0
litellm>=1.58.0
pytest>=8.4.1
ruff>=0.12.8
httpx>=0.28.1
